Create java006
The runs scored by N batsmen of a cricket team is passed as the input to the program. The program must print the name of the batsman who scored the highest runs. (You can assume that no two batsmen will be the top scorers).

nput:

5

BatsmanA,45

BatsmanB,52

BatsmanC, 12

BatsmanD,9

BatsmanE,78

Output:

BatsmanE