java008
A certain number of people attended a meeting which was to begin at 10:00 am on a given day. The arrival time in HH:MM format of those who attended the meeting is passed as the input in a single line, with each arrival time by a space. The program must print the count of people who came late (after 10:00 am) to the meeting.

Input Format: The first line contains the arrival time separated by a space.

Output Format: The first line contains the count of late comers.
Input: 10:00 9:55 10:02 9:45 11:00

Output: 2